Changed some macros to explicitly specify the integer types.
This was causing a compile error. As I compiled using a new version
of the gcc cross-compiler, it was implicitly assuming signed integer
types where I assumed the unsigned 64-bit type. Now the raw integers
are explicitly typed to "size_t", the u64 type.

I should be careful in the future to explicitly specify this type.